requirements.txt
streamlit>=1.28
langchain>=0.0.217
wikipedia-api>=0.5.4
duckduckgo-search
trubrics>=1.4.3
streamlit-feedback
